Fix Context Debugger Startup
