+++ client/src/services/context.ts
@@ -1,6 +1,6 @@
 import { extension } from "../state";
 import { LJContext, Range, LJVariable } from "../types/context";
-import { SourcePosition } from "../types/diagnostics";
+import { RefinementMismatchError, SourcePosition } from "../types/diagnostics";
 import { getOriginalVariableName } from "../utils/utils";
 
 export function handleContext(context: LJContext) {
@@ -11,6 +11,7 @@ export function handleContext(context: LJContext) {
     const { allVars, visibleVars } = getSelectionContextVariables(extension.file, extension.currentSelection);
     extension.context.visibleVars = visibleVars;
     extension.context.allVars = allVars;
+    updateErrorAtCursor();
     extension.webview.sendMessage({ type: "context", context: extension.context, errorAtCursor: extension.errorAtCursor });
 }
 
@@ -22,6 +23,22 @@ export function getSelectionContextVariables(file: string, selection: Range): { 
     const visibleVarsByAnnotationPosition = getVisibleVariables(variablesInScope, file, selection, true);
     const allVars = sortVariables(normalizeVariableRefinements([...globalVars, ...visibleVarsByPosition]));
     return { visibleVars: visibleVarsByAnnotationPosition, allVars };
+}
+
+export function updateErrorAtCursor() {
+    if (!extension.file || !extension.currentSelection) return;
+    const errors: RefinementMismatchError[] = extension.diagnostics?.filter(d => d.type === 'refinement-error' || d.type === 'state-refinement-error') as RefinementMismatchError[] || [];
+    const scopes = extension.context?.fileScopes[extension.file] || [];
+    const errorAtCursor = errors.find(error => {
+        if (!error.position) return false;
+        const sameFile = error.position.file === extension.file;
+        const beforeCursor = isPositionBefore(error.position, extension.currentSelection);
+        if (!sameFile || !beforeCursor) return false;
+        // check if error is within a scope that contains the cursor
+        const errorScope = scopes.find(scope => isRangeWithin(error.position, scope));
+        return errorScope && isRangeWithin(extension.currentSelection, errorScope);
+    });
+    extension.errorAtCursor = errorAtCursor;
 }
 
 function getVariablesInScope(variables: LJVariable[], file: string, selection: Range): LJVariable[] {
